rectifying faults from moving folders
