+++ Server/server.js
@@ -8,7 +8,7 @@ server.use(cors())
 
 const port = 3000;
 
-const googleRoutes = require('./models/controllers/controllers');
+const googleRoutes = require('./controllers/controllers');
 server.use('/google', googleRoutes);
 
 server.listen(port, () => console.log(`Server running from http://localhost:${port}`))
